% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzposterior.R
\name{calc}
\alias{calc}
\title{Calculate statistics}
\usage{
calc(x, f, ...)
}
\arguments{
\item{x}{an \code{inzposterior} object}

\item{f}{a function to compute for each parameter}

\item{...}{arguments passed to \code{f}}
}
\value{
a vector or matrix of values
}
\description{
Calculate statistics
}
