% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_posterior.R
\name{calculate_posterior}
\alias{calculate_posterior}
\alias{calculate_posterior.inz_dbeta}
\alias{calculate_posterior.inz_ddir}
\alias{calculate_posterior.inz_dNIG}
\alias{calculate_posterior.inz_dNIG_reg}
\title{Calculate the Posterior Distribution}
\usage{
calculate_posterior(prior, ...)

\method{calculate_posterior}{inz_dbeta}(prior, cred_level = 95, signif_value = 4, ...)

\method{calculate_posterior}{inz_ddir}(prior, cred_level = 95, signif_value = 4, ...)

\method{calculate_posterior}{inz_dNIG}(prior, cred_level = 95, signif_value = 4, ...)

\method{calculate_posterior}{inz_dNIG_reg}(prior, cred_level = 95, signif_value = 4, ...)
}
\arguments{
\item{prior}{The prior object.}

\item{...}{Currently no additional arguments.}

\item{cred_level}{The credible level (\%) for the credible interval (default of 95).}

\item{signif_value}{The number of significant figures for output (default of 4).}
}
\value{
Returns an object of the same \link{class} as the \code{prior}.
It is a list which contains the likelihood, the prior, and the posterior
parameter values.

The \code{cred_level} and \code{signif_value} is stored as an attribute.
}
\description{
\code{calculate_posterior} is a generic function used to calculate the posterior
distribution given the prior and its associated likelihood.
}
\details{
\bold{Beta-Binomial conjugacy:}

The posterior parameters for \code{inz_dbeta} are calculated and updated using:
\deqn{
\begin{aligned}
\alpha_{post} &= \alpha_{prior} + x \\
\beta_{post} &= \beta_{prior} + N - x
\end{aligned}
}

\bold{Dirichlet-Multinomial conjugacy:}

The posterior parameters for \code{inz_ddir} are calculated and updated using:
\deqn{\boldsymbol{\alpha}_{post} = \boldsymbol{\alpha}_{prior} + \mathbf{x}}

\bold{Normal-Inverse-Gamma Prior, Normal Likelihood conjugacy:}

The posterior parameters for \code{inz_dNIG} are calculated and updated using:
\deqn{
\begin{aligned}
V_n &= \left( \frac{1}{V_0} + n \right)^{-1} \\
m_n &= V_n \left( \frac{m_0}{V_0} + n \bar{x} \right) \\
a_n &= a_0 + \frac{n}{2} \\
b_n &= b_0 + \frac{1}{2} \left(\frac{m_0^2}{V_0} + \sum_{i} x_i^2 - \frac{m_n^2}{V_n} \right)
\end{aligned}
}
(Murphy, 2007, Section 6.3).

\bold{Normal-Inverse-Gamma Prior, Normal Likelihood conjugacy for regression:}

The posterior parameters for \code{inz_dNIG_reg} is calculated and updated using:
\deqn{
\begin{aligned}
\boldsymbol{\Lambda}_n &= \mathbf{X}^T \mathbf{X} + \boldsymbol{\Lambda}_0 \\
\boldsymbol{\mu}_n &= \boldsymbol{\Lambda}_n^{-1} (\mathbf{X}^T \mathbf{X} \hat{\boldsymbol{\beta}} + \boldsymbol{\Lambda}_0 \boldsymbol{\mu}_0) \\
a_n &= a_0 + \frac{n}{2} \\
b_n &= b_0 + \frac{1}{2} (\mathbf{y}^T \mathbf{y} + \boldsymbol{\mu}_0^T \boldsymbol{\Lambda}_0 \boldsymbol{\mu}_0 - \boldsymbol{\mu}_n^T \boldsymbol{\Lambda}_n \boldsymbol{\mu}_n)
\end{aligned}
}
}
\examples{
# Beta-Binomial example
lik <- inz_lbinom(surf_data, Gender)
prior <- inz_dbeta(likelihood = lik)
calculate_posterior(prior = prior)

# Regression example (Normal-Inverse-Gamma)
lik <- inz_lnorm(surf_data, Income, Hours)
prior <- inz_dNIG(likelihood = lik)
calculate_posterior(prior = prior)

}
\references{
Murphy, K. P. (2007). \emph{Conjugate Bayesian analysis of the Gaussian distribution} (Technical Report).
The University of British Columbia. \url{https://www.cs.ubc.ca/~murphyk/Papers/bayesGauss.pdf}
}
