% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lm.R
\name{estimate_lm}
\alias{estimate_lm}
\alias{estimate_lm.default}
\alias{estimate_lm.formula}
\title{Estimate a linear regression model}
\usage{
estimate_lm(y, ...)

\method{estimate_lm}{default}(y, x, data, ...)

\method{estimate_lm}{formula}(y, data, ...)
}
\arguments{
\item{y}{a vector of response values or a formula}

\item{...}{additional arguments passed to child functions}

\item{x}{predictor(s), a matrix}

\item{data}{a data.frame containing the data}
}
\value{
a posterior distribution for the regression parameters
}
\description{
Estimate a linear regression model
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method

\item \code{formula}: Method for a formula
}}

\examples{
post <- estimate_lm(Petal.Length ~ Petal.Width, data = iris)
summary(post)
plot(post)
}
\seealso{
\link{inzposterior}
}
