% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{estimate_proportions}
\alias{estimate_proportions}
\title{Exact posterior for proportions of one categorical variable}
\usage{
estimate_proportions(x, alpha = rep(1, length(x)))
}
\arguments{
\item{x}{a vector of counts ("successes" in each group)}

\item{alpha}{a vector of prior concentration parameters (default uniform)}
}
\value{
the posterior distribution estimate
}
\description{
Exact posterior estimation for proportions of one categorical variable
using dirichlet prior and multinomial likelihood.
}
\details{
Assumes multinomial likelihood and Dirichlet prior
}
\examples{
\dontrun{
x = c(100, 200, 300)
post = estimate_proportions(x)
}
}
