% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_lm}
\alias{gibbs_lm}
\title{Gibbs Sample for Linear Regression}
\usage{
gibbs_lm(y, x, steps = 1000L)
}
\arguments{
\item{y}{a vector of response values}

\item{x}{a matrix of explanatory variables}
}
\description{
Gibbs sampler for linear regression
}
