% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_lm.R
\name{gibbs_lm_R}
\alias{gibbs_lm_R}
\title{Gibbs sampler for linear regression}
\usage{
gibbs_lm_R(y, X, steps)
}
\arguments{
\item{y}{a vector of the response variable}

\item{X}{a matrix whose columns are the explanatory variables}

\item{steps}{number of iterations to run Gibbs sampler for}
}
\value{
a list of posterior samples for regression coefficients
}
\description{
Gibbs sampler for linear regression
}
\details{
Assumes Normal likelihood and joint prior proportional to 1 / sigma2
}
\examples{
\dontrun{
set.seed(1234)
n = 100
sigma = 40
beta = c(150, 5, 10)
x = seq(-4, 10, length = n)
y = beta[1] + beta[2] * x + beta[3] * x ^ 2 + rnorm(n, 0, sigma)
X = cbind(x, x ^ 2)
mcmc = gibbs_lm_R(y, X, 10000)
}
}
