% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{gibbs_one_categorical}
\alias{gibbs_one_categorical}
\title{Gibbs sampler to sample proportions of one categorical variable}
\usage{
gibbs_one_categorical(x, steps, alpha = rep(1, length(x)))
}
\arguments{
\item{x}{a vector of counts ("successes" in each group)}

\item{steps}{number of iterations to run Gibbs sampler for}

\item{alpha}{a vector of prior concentration parameters (default uniform)}
}
\value{
a list of posterior samples for the proportions of one categorical variable
}
\description{
Gibbs sampler to sample proportions of one categorical variable
}
\details{
Assumes multinomial likelihood and Dirichlet prior
}
\examples{
\dontrun{
set.seed(1234)
x = c(100, 200, 300)
mcmc = gibbs_one_categorical(x, 20000)
}
}
