% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_one_numeric.R
\name{gibbs_one_numeric}
\alias{gibbs_one_numeric}
\title{Gibbs sampler to sample mean and variance of one numeric variable}
\usage{
gibbs_one_numeric(
  y,
  steps,
  burnin,
  thin = 1,
  mu.0 = 0,
  sigma2.0 = 1e+06,
  alpha = 0.001,
  beta = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{y}{a vector of values}

\item{steps}{number of iterations to run Gibbs sampler for}

\item{burnin}{number of burn-in iterations to discard before proper steps}

\item{thin}{thinning factor (default 1)}

\item{mu.0}{prior mean for mu (default 0)}

\item{sigma2.0}{prior variance for mu (default 1e6)}

\item{alpha}{prior shape parameter for sigma2 (default 1e-3)}

\item{beta}{prior scale parameter for sigma2 (default 1e-3)}

\item{verbose}{logical, if \code{TRUE} progress will be printed to console}
}
\value{
a list of posterior samples for the mean and variance of one numeric variable
}
\description{
Gibbs sampler to sample mean and variance of one numeric variable
}
\details{
Assumes conjugate Normal-Inverse-Gamma priors on mean and variance
}
\examples{
\dontrun{
set.seed(1234)
n = 1000
y = rnorm(n, 10, sqrt(4))
mcmc = gibbs_one_numeric(y, 20000, 5000)
}
}
