% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_conj_prior.R
\name{inz_dbeta}
\alias{inz_dbeta}
\title{Beta Distribution for Prior and Posterior}
\usage{
inz_dbeta(likelihood = NULL, alpha = NULL, beta = NULL)
}
\arguments{
\item{likelihood}{The likelihood object of class \code{"inz_lbinom"}.}

\item{alpha}{The shape parameter \eqn{\alpha}.}

\item{beta}{The shape parameter \eqn{\beta}.}
}
\value{
Returns an object of \link{class} \code{"inz_dbeta"}.

An object of class \code{"inz_dbeta"} is a list which contains the following:
\item{alpha}{the shape parameter \eqn{\alpha}.}
\item{beta}{the shape parameter \eqn{\beta}.}

If the function is used to construct the prior, the \code{likelihood} is stored
as an attribute.
}
\description{
Creates an object containing the parameters of a Beta distribution.
This is used for constructing the prior or storing the information
of the posterior.
}
\details{
\itemize{
\item Parameter 1: shape \eqn{\alpha > 0}
\item Parameter 2: shape \eqn{\beta > 0}
}

The Beta distribution is the conjugate prior for (but is not limited to)
the Binomial likelihood.

\bold{For prior use only:}
If no \code{alpha} or \code{beta} parameter values are provided, a default
value of \code{\eqn{\alpha}=1} and \code{\eqn{\beta}=1} will be used respectively.
Hence, a default prior of Beta(1,1) will be used if \code{alpha} and \code{beta}
are both \code{NULL}, which is equivalent to a uninformative, Uniform(0,1) prior.
}
\examples{
# Constructing the prior with the likelihood (default prior is used)
lik <- inz_lbinom(surf_data, Gender)
inz_dbeta(likelihood = lik)

# Using a subjective prior 
inz_dbeta(likelihood = lik, alpha = 8, beta = 2)

# Grouped data example
grouped_lik <- inz_lbinom(surf_data, Gender, Qualification)
inz_dbeta(likelihood = grouped_lik)

# Example of inz_dbeta usage in the calculate_posterior function
inz_dbeta(alpha = 108, beta = 94)
}
\seealso{
\code{\link{inz_lbinom}}

\code{\link{calculate_posterior}}
}
