% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_conj_prior.R
\name{inz_ddir}
\alias{inz_ddir}
\title{Dirichlet Distribution for Prior and Posterior}
\usage{
inz_ddir(likelihood = NULL, alpha = NULL, k = NULL)
}
\arguments{
\item{likelihood}{The likelihood object of class \code{"inz_lmulti"}.}

\item{alpha}{The shape parameters \eqn{\alpha} (vector or matrix).}

\item{k}{The number of categories.}
}
\value{
Returns an object of \link{class} \code{"inz_ddir"}.

An object of class \code{"inz_ddir"} is a list which contains the following:
\item{alpha}{a vector or a matrix of the shape parameters \eqn{\alpha}.}
\item{k}{the number of categories.}

If the function is used to construct the prior, the \code{likelihood} is stored
as an attribute.
}
\description{
Creates an object containing the parameters of a Dirichlet distribution.
This is used for constructing the prior or storing the information
of the posterior.
}
\details{
\itemize{
\item The shape parameter: \eqn{\alpha_i > 0}; each corresponding to a level/group
in the categorical variable
\item \eqn{k ≥ 2}
}

The Dirichlet distribution is the conjugate prior for (but is not limited to)
the Multinomial likelihood.

If the data is grouped, \code{alpha} should be a matrix with one row per group.
Otherwise, \code{alpha} should be a vector.

\bold{For prior use only:}
If no \code{alpha} values are provided, the default value of \eqn{\alpha_i = 1},
will be used for the prior (e.g. Dirichlet(1,1,1,1) for a k=4 level categorical
variable).
}
\examples{
# Constructing the prior with the likelihood (default prior is used)
lik <- inz_lmulti(surf_data, Qualification)
inz_ddir(likelihood = lik)

# Using a subjective prior (prior belief that degree and school qualifications are more common)
inz_ddir(likelihood = lik, alpha = c(10, 2, 10, 2))

# Grouped data example
grouped_lik <- inz_lmulti(surf_data, Qualification, Ethnicity)
inz_ddir(likelihood = grouped_lik)

# Example of inz_ddir usage in the calculate_posterior function
posterior <- inz_ddir(alpha = c(29,40,67,68), k = 4)
}
\seealso{
\code{\link{inz_lmulti}}

\code{\link{calculate_posterior}}
}
