% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_likelihood.R
\name{inz_lbinom}
\alias{inz_lbinom}
\title{Construct a Likelihood Object for Binomial Data}
\usage{
inz_lbinom(data, primary, secondary = NULL)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{primary}{The primary binary variable of interest.}

\item{secondary}{An optional secondary variable of interest (for grouping).}
}
\value{
Returns an object of \link{class} \code{"inz_lbinom"}.

An object of class \code{"inz_lbinom"} is a list which contains the following:
\item{x}{the number of successes.}
\item{N}{the total number of observations.}
\item{levels}{the levels of the primary variable (the binary outcomes).}
\item{groups}{the levels of the secondary variable, or "none" if not grouped.}
}
\description{
Creates a likelihood object from a binary categorical variable,
grouped by a secondary categorical variable if provided.
}
\details{
Both \code{primary} and \code{secondary} must be categorical variables.
They can be passed as a string (e.g. \code{"Gender"}) or as a non-string (e.g. \code{Gender}).

The \code{primary} variable must be binary (i.e. a two-level categorical variable).
The first level of the variable is treated as the 'success' (\code{x}).

For grouped data, the function calculates the Binomial parameters, \code{x} and \code{N},
for each level of the \code{secondary} variable.
}
\examples{
# Single variable
inz_lbinom(surf_data, Gender)

# Two variables (grouped data)
inz_lbinom(surf_data, Gender, Qualification)
}
