% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_likelihood.R
\name{inz_lmulti}
\alias{inz_lmulti}
\title{Construct a Likelihood Object for Multinomial Data}
\usage{
inz_lmulti(data, primary, secondary = NULL)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{primary}{The primary variable of interest.}

\item{secondary}{An optional secondary variable of interest (for grouping).}
}
\value{
Returns an object of \link{class} "inz_lmulti".

An object of class "inz_lmulti" is a list which contains the following:
\item{x}{the count of observations in each level (as a vector or as a table if grouped).}
\item{levels}{the levels of the primary variable.}
\item{groups}{the levels of the secondary variable, or "none" if not grouped.}
}
\description{
Creates a likelihood object from a categorical variable with three or more levels,
grouped by a secondary categorical variable if provided.
}
\details{
Both \code{primary} and \code{secondary} must be categorical variables.
They can be passed as a string (e.g. \code{"Qualification"}) or as a non-string (e.g. \code{Qualification}).

The \code{primary} variable must have three or more levels.
If the \code{primary} variable has only two levels, use \code{\link{inz_lbinom}}.
}
\examples{
# Single variable
\dontrun{inz_lmulti(surf_data, Qualification)}

# Two variables (grouped data)
\dontrun{inz_lmulti(surf_data, Qualification, Gender)}
\dontrun{inz_lmulti(surf_data, Qualification, Ethnicity)}
}
\seealso{
\code{\link{inz_lbinom}}
}
