% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_likelihood.R
\name{inz_lnorm}
\alias{inz_lnorm}
\title{Construct a Likelihood Object for Normal Data}
\usage{
inz_lnorm(data, primary, secondary = NULL)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{primary}{The primary numeric variable of interest.}

\item{secondary}{An optional secondary variable of interest (categorical for grouping
or numeric for regression).}
}
\value{
Returns an object of \link{class} \code{"inz_lnorm"}.

An object of class \code{"inz_lnorm"} is a list which contains the following:

\bold{For a single numeric variable:}
\item{x}{the numeric variable.}
\item{summary_stat}{the summary statistics of the variable.}
\item{x_bar}{the sample mean.}
\item{n}{the sample size.}
\item{raw_sum_sq}{the raw sum of squares.}
\item{primary_var}{the name of the numeric variable.}

\bold{For numeric and categorical:}
\item{grouped_data}{the numeric variable grouped by the categorical variable.}
\item{summary_stat}{the summary statistics for each group.}
\item{x_bar}{a vector of sample means for each group.}
\item{n}{a vector of sample sizes for each group.}
\item{raw_sum_sq}{a vector of the raw sum of squares for each group.}
\item{primary_var}{the name of the numeric variable.}
\item{secondary_var}{the name of the categorical variable.}

\bold{For numeric and numeric:}
\item{x}{the explanatory/independent variable.}
\item{y}{the response/dependent variable.}
\item{spearman_correlation}{Spearman's rank correlation coefficient.}
\item{x_label}{the name of the explanatory variable.}
\item{y_label}{the name of the response variable.}
}
\description{
Creates a likelihood object from a numeric variable, optionally grouped by a
categorical variable. Or, optionally use another numeric variable to explore
the relationship.
}
\details{
\code{primary} and \code{secondary} arguments can be passed as a string
(e.g. \code{"Height"}) or as a non-string (e.g. \code{Height}).

The function behaves differently depending on the input.

\strong{Single variable - numeric:} Computes summary statistics for the
numeric variable.

\strong{Two variables - numeric and categorical:} Groups the numeric variable
by the categorical variable and computes summary statistics for each group.
The function automatically identifies the variable which is numeric and uses it
as the \code{primary} variable regardless of the order in which the variables
are inputted.

\strong{Two variables - numeric and numeric:} Computes Spearman's correlation.
The \code{secondary} variable is used as the explanatory variable and the
\code{primary} variable is used as the response variable for regression.
}
\examples{
# Single variable
\dontrun{inz_lnorm(surf_data, Hours)}

# Numeric and Categorical (grouped data)
\dontrun{inz_lnorm(surf_data, Income, Qualification)}
\dontrun{inz_lnorm(surf_data, Qualification, Income)} # gives the same output

# Numeric and Numeric (regression)
\dontrun{inz_lnorm(surf_data, Income, Hours)}
}
