% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightBayes.R
\docType{package}
\name{iNZightBayes}
\alias{iNZightBayes-package}
\alias{iNZightBayes}
\title{iNZightBayes}
\description{
Description of your package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayes.inzight.nz}
  \item Report bugs at \url{https://github.com/iNZightVIT/iNZightBayes/issues}
}

}
\author{
Tom Elliott \href{mailto:tom.elliott@auckland.ac.nz}{tom.elliott@auckland.ac.nz}
}
