% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzposterior.R
\name{inzposterior}
\alias{inzposterior}
\alias{print.inzposterior}
\alias{summary.inzposterior}
\alias{summary.inzexact}
\alias{print.inzexactsummary}
\alias{plot.inzposterior}
\alias{plot.inzexact}
\alias{mean.inzposterior}
\title{iNZight Posterior Object}
\usage{
\method{print}{inzposterior}(x, ...)

\method{summary}{inzposterior}(object, ...)

\method{summary}{inzexact}(object, ...)

\method{print}{inzexactsummary}(x, ...)

\method{plot}{inzposterior}(x, y, ...)

\method{plot}{inzexact}(x, y, ...)

\method{mean}{inzposterior}(x, ...)
}
\arguments{
\item{x}{an object of class \code{inzposterior}}

\item{...}{additional arguments passed to methods}

\item{object}{an object of class \code{inzposterior}}

\item{y}{optional, the parameter to plot. All shown if ommitted}
}
\description{
All posterior results are returned with this class, which includes
methods for obtaining posterior summaries and samples.
}
\section{Functions}{
\itemize{
\item \code{print.inzposterior}: Print method for iNZight posterior obejcts

\item \code{summary.inzposterior}: Summary method for iNZight posterior objects

\item \code{summary.inzexact}: Summary method for iNZight posterior objects

\item \code{print.inzexactsummary}: Plot method for iNZight posterior objects

\item \code{plot.inzposterior}: Plot method for iNZight posterior objects

\item \code{plot.inzexact}: Plot method for iNZight exact-posterior objects

\item \code{mean.inzposterior}: Returns the mean of the estimated parameters
}}

