% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inz_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\alias{summary.inz_dbeta}
\alias{summary.inz_ddir}
\alias{summary.inz_dNIG}
\alias{summary.inz_dNIG_reg}
\title{Posterior Object Summary}
\usage{
\method{summary}{inz_dbeta}(posterior_obj, ...)

\method{summary}{inz_ddir}(posterior_obj, ...)

\method{summary}{inz_dNIG}(posterior_obj, ...)

\method{summary}{inz_dNIG_reg}(posterior_obj, ...)
}
\arguments{
\item{posterior_obj}{A posterior object returned by \code{calculate_posterior}.}

\item{...}{Currently no additional arguments.}
}
\value{
An object of class \code{summary.inz_*}, which is used by the corresponding
\code{print} method to automatically print the output.
}
\description{
Compute and display posterior summaries. This includes the point
estimate (posterior mean) and equal-tailed credible intervals.
}
\details{
\bold{Beta-Binomial} (\code{inz_dbeta}):

The point estimates are calculated using the expectation formula of the Beta
distribution:
\deqn{\frac{\alpha}{\alpha+\beta}}

The credible intervals are calculated using the quantile function \link{qbeta}.

\bold{Dirichlet-Multinomial} (\code{inz_ddir}):

The point estimates are calculated using the expectation formula of the Dirichlet
distribution:
\deqn{\frac{\alpha_i}{\alpha_0}}

The marginal distribution of Dirichlet follows a Beta distribution.
Hence, the credible intervals are calculated using the quantile function
\link{qbeta}.

\bold{Normal-Inverse-Gamma Prior, Normal Likelihood} (\code{inz_dNIG} or \code{inz_dNIG_reg}):

The posterior mean computed in \code{calculate_posterior}
(\eqn{m_n} or \eqn{\boldsymbol{\mu}_n} for regression) are the point estimates.

The marginal posterior distribution of the mean follows a t-distribution.
Hence, the credible intervals are calculated using the quantile function \link{qt}.
}
\examples{
# Beta-Binomial example
lik <- inz_lbinom(surf_data, Gender)
prior <- inz_dbeta(likelihood = lik)
posterior <- calculate_posterior(prior = prior)
summary(posterior)

# Regression example (Normal-Inverse-Gamma)
lik <- inz_lnorm(surf_data, Income, Hours)
prior <- inz_dNIG(likelihood = lik)
posterior <- calculate_posterior(prior = prior)
summary(posterior)

}
\seealso{
\code{\link{calculate_posterior}}
}
