% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{rdirichlet}
\alias{rdirichlet}
\title{The Dirichlet Distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{number of random samples}

\item{alpha}{vector of concentration parameters}
}
\value{
a matrix of random samples from Diriclet distribution
}
\description{
Random generation of Dirichlet distribution
}
\details{
Draws independent Gamma random variables and normalises
}
\examples{
\dontrun{
set.seed(1234)
rdirichlet(10, c(1, 1, 1))
}
}
