% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{estimate_mean}
\alias{estimate_mean}
\title{Estimate the sample mean}
\usage{
estimate_mean(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a posterior distribution for the mean
}
\description{
Estimate the sample mean
}
\examples{
estimate_mean(rnorm(100, 50, 5))
}
