% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_one_numeric_cpp}
\alias{gibbs_one_numeric_cpp}
\title{Gibbs sampler for single numeric mean and variance}
\usage{
gibbs_one_numeric_cpp(
  y,
  steps,
  burnin,
  thin = 1L,
  mu_0 = 0,
  sigma2_0 = 1e+06,
  alpha = 0.001,
  beta = 0.001
)
}
\arguments{
\item{y}{vector of values}

\item{steps}{number of steps}
}
\description{
Description and details
}
