% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightBayes.R
\docType{package}
\name{iNZightBayes}
\alias{iNZightBayes}
\title{iNZightBayes}
\description{
Description of your package
}
\author{
Tom Elliott \href{mailto:tom.elliott@auckland.ac.nz}{tom.elliott@auckland.ac.nz}
}
