% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{inzposterior}
\alias{inzposterior}
\alias{print.inzposterior}
\alias{summary.inzposterior}
\alias{plot.inzposterior}
\title{iNZight Posterior Object}
\usage{
\method{print}{inzposterior}(x, ...)

\method{summary}{inzposterior}(object, ...)

\method{plot}{inzposterior}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{inzposterior}}

\item{...}{additional arguments passed to methods}

\item{object}{an object of class \code{inzposterior}}

\item{y}{optional, the parameter to plot. All shown if ommitted}
}
\description{
iNZight Posterior Object
}
\section{Functions}{
\itemize{
\item \code{print.inzposterior}: Print method for iNZight posterior obejcts

\item \code{summary.inzposterior}: Summary method for iNZight posterior objects

\item \code{plot.inzposterior}: Plot method for iNZight posterior objects
}}

